+++ CHANGELOG.md
@@ -1,5 +1,37 @@
 # Changelog
 
+## [2.9.0](https://github.com/erxes/erxes/compare/2.8.3...2.9.0) (2025-04-02)
+
+### Features
+
+* add phone number and email validation in Call component ([b16e93a](https://github.com/erxes/erxes/commit/b16e93a3d2d09074dc3407cc85e65acb3481649d))
+* **cms:** Implement custom fields group functionality for CMS categories and pages, extending the existing custom fields ([#5987](https://github.com/erxes/erxes/issues/5987)) ([f82d136](https://github.com/erxes/erxes/commit/f82d136a73e67ca610d61f6863531b5142df7627))
+* **pms:** check-in & check-out ([9d89279](https://github.com/erxes/erxes/commit/9d8927908175e931a63a07bf6b4433764034909f))
+
+### Bug Fixes
+
+* **bm:** sort on tour,itinerary, tour group ([d0ae651](https://github.com/erxes/erxes/commit/d0ae6512095dafea81f676ac23971cc903bd2987))
+* **cards:** null pointer on chart view ([bfb1a8c](https://github.com/erxes/erxes/commit/bfb1a8c1512315b7d787a75b271b56b57866d83a))
+* **core:** can't import unitPrice is string on products ([4cbeb4b](https://github.com/erxes/erxes/commit/4cbeb4b2d31f4f0026d58f7321bbeb387be732f7))
+* improve 2-way audio for multiple operators ([97940c1](https://github.com/erxes/erxes/commit/97940c170564ac3087430906a771644014e68695))
+* **pms:** pms rooms pagination ([6cef89b](https://github.com/erxes/erxes/commit/6cef89bd40471c78bc506d0801b62f3cf8b612b8))
+* **tms:** tour images and order fields ([234adb1](https://github.com/erxes/erxes/commit/234adb1403c13e85fa27cc3d4822a253868c7948))
+* yaml files ([858aaf5](https://github.com/erxes/erxes/commit/858aaf54707aff14f40981dc45092e4b29809656))
+## [2.9.0-rc.0](https://github.com/erxes/erxes/compare/2.8.3...2.9.0) (2025-03-16)
+
+### Features
+
+* Accountings improve ([#5966](https://github.com/erxes/erxes/issues/5966)) ([f7ef547](https://github.com/erxes/erxes/commit/f7ef5470f4afbcb5aed5f2044456f8642825b9a9))
+* improve sales-ui, improve loans, add plugin loansresearch ([#5968](https://github.com/erxes/erxes/issues/5968)) ([19e5397](https://github.com/erxes/erxes/commit/19e539708a082b40d7a6fb8ce52ae982fd0c7a93)), closes [#5895](https://github.com/erxes/erxes/issues/5895) [#5898](https://github.com/erxes/erxes/issues/5898)
+
+### Bug Fixes
+
+* **pms:** room change permission ([442c512](https://github.com/erxes/erxes/commit/442c51245ead4c39851ebdbbe5e7154be9392203))
+* **pms:** room change permission ([20079a9](https://github.com/erxes/erxes/commit/20079a911276ea10a1ea0bdd5196c27a93eb4739))
+* **tms:** extra ([77cc177](https://github.com/erxes/erxes/commit/77cc177515d513823aa33f1e42ff819fde2cf27d))
+* **tms:** extrafields ([8e4ecfb](https://github.com/erxes/erxes/commit/8e4ecfb0958a5ec06aff2db7e68301e401dcccc1))
+
+## [2.9.0-rc.0](https://github.com/erxes/erxes/compare/2.8.0...2.9.0-rc.0) (2025-03-16)
 ## [2.8.3](https://github.com/erxes/erxes/compare/2.8.2...2.8.3) (2025-03-26)
 
 ### Bug Fixes